refactor: rename interfaces
