+++ src/governance.ts
@@ -3,10 +3,10 @@ import {GovernanceService, idlFactory} from '../candid/governance.idl';
 import {idlFactory as certifiedIdlFactory} from '../candid/governance.certified.idl';
 import {RequestConverters} from './canisters/governance/request.converters';
 import {ResponseConverters} from './canisters/governance/response.converters';
-import {KnownNeuron, ListProposalsRequest, ListProposalsResponse,} from './types/governance';
+import {KnownNeuron, ListProposalsRequest, ListProposalsResponse,} from './types/governance_converters';
 import {MAINNET_GOVERNANCE_CANISTER_ID} from './constants/canister_ids';
 import {defaultAgent} from './utils/agent.utils';
-import {GovernanceCanisterOptions} from './types/governance_canister';
+import {GovernanceCanisterOptions} from './types/governance';
 
 export class GovernanceCanister {
   private constructor(
